+++ R/OverExpectedHiC.R
@@ -5,6 +5,8 @@
 #'  computed per genomic distance.
 #' @param hicLst <List[ContactMatrix][InteractionSet::ContactMatrix()]>:
 #'  The HiC maps list.
+#' @param method <character> Options are "mean_non_zero", "mean_total",
+#'  or "lieberman". Look at details for more (Default: "mean_non_zero")
 #' @param cores <numerical> : Number of cores to be used. (Default 1)
 #' @param verbose <logical>: If TRUE show the progression in console.
 #'  (Default FALSE)
@@ -12,15 +14,29 @@
 #'  per chromosome ("per_seq"), all chromosomes ("total") or not (NULL). 
 #'  (Default "per_seq")
 #' @return A matrices list.
+#' @export
 #' @importFrom rlang .data
 #' @import ggplot2
+#' @importFrom reshape melt
+#' @details
+#' Methods to calculate expected values per distance:
+#' \itemize{
+#' \item "mean_non_zero": for each distance, average contact value is
+#' calculated using only non-zero values.
+#' \item "mean_total": for each distance, average contact value is
+#' calculated using all values at this distance.
+#' \item "lieberman": for each distance, contact values are summed and
+#' divided by chromsome length minus distance. Only for cis contacts.
+#' }
 #' @examples
 #' # Note: run HicAggR::BalanceHiC before OverExpectedHiC calculation.
 #' data(HiC_Ctrl.cmx_lst)
 #' OverExpectedHiC(HiC_Ctrl.cmx_lst)
 #'
 OverExpectedHiC <- function(
-    hicLst, verbose = FALSE, cores = 1, plot_contact_vs_dist="per_seq"
+    hicLst, method = "mean_non_zero",
+    verbose = FALSE, cores = 1, 
+    plot_contact_vs_dist="per_seq"
 ) {
     resolution.num <- attributes(hicLst)$resolution
     matricesNames.chr <- names(hicLst)
@@ -34,15 +50,48 @@ OverExpectedHiC <- function(
             matrixName.chr <- matricesNames.chr[matrixName.ndx]
             if (hicLst[[matrixName.chr]]@metadata$type == "cis") {
                 hic.dtf <- MeltSpm(hicLst[[matrixName.chr]]@matrix)
+                if(method != "mean_non_zero"){
+                    hic_dtf_total <- 
+                    as.matrix(hicLst[[matrixName.chr]]@matrix)|>
+                    reshape::melt()|>
+                    `colnames<-`(c("i","j","x"))|>
+                    dplyr::filter(.data$i <= .data$j) |>
+                    dplyr::mutate(distance =
+                        1 + (.data$j - .data$i) * resolution.num)|>
+                        dplyr::select(c("x", "distance"))
+                }
                 hic.dtf <- dplyr::mutate(
                     hic.dtf,
-                    distance = 1 + (hic.dtf$j - hic.dtf$i) * resolution.num) |>
+                    distance = 
+                        (1 + (hic.dtf$j - hic.dtf$i) * resolution.num)) |>
                     dplyr::select(c("x", "distance"))
+                if(method == "lieberman"){
+                    chr_length <- attributes(hicLst)$chromSize$length[which(
+                        attributes(hicLst)$chromSize$name == 
+                        unlist(strsplit(matrixName.chr,split = "_"))[1])]
+
+                    expected.dtf <- dplyr::group_by(
+                        hic_dtf_total, distance = hic_dtf_total$distance) |>
+                        # get distance, chr_length and contacts!!!
+                        dplyr::summarise_at(.vars="x",
+                            .funs = list(total=sum)) |>
+                        dplyr::mutate(factor=chr_length - .data$distance) |>
+                        dplyr::mutate(expected=.data$total/.data$factor) |>
+                        dplyr::select(-c("factor","total"))
+
+                    }else if(method == "mean_total"){
+                    expected.dtf <- dplyr::group_by(
+                        hic_dtf_total, distance = hic_dtf_total$distance) |>
+                        dplyr::summarise_at(
+                        .vars = "x", .funs = list(expected = mean)
+                        )
+                    }else{
                 expected.dtf <- dplyr::group_by(
                     hic.dtf, distance = hic.dtf$distance) |>
                     dplyr::summarise_at(
                         .vars = "x", .funs = list(expected = mean)
                     )
+                    }
                 expected.num <- dplyr::left_join(
                     hic.dtf,
                     expected.dtf,
@@ -74,7 +123,7 @@ OverExpectedHiC <- function(
     expected.dtf <- expected.lst[cisNames.chr] |>
         do.call(what = rbind)
     expected.lst <- expected.lst[cisNames.chr]
-    
+
     if(plot_contact_vs_dist=="per_seq" && length(expected.lst)>0){
         data <- do.call(rbind,
             lapply(