+++ R/compare_to_background.R
@@ -1,31 +1,51 @@
 
 #' Computes z.test for each target couple over background couples.
 #'
-#' @param hicList <List[ContactMatrix][InteractionSet::ContactMatrix()]>: 
+#' @param hicList <List[ContactMatrix][InteractionSet::ContactMatrix()]>:
 #' The HiC maps list.
 #' @param matrices <list[matrix]>: The matrices list.
-#' @param genomicConstraint <GRanges>: GRanges object of constraint regions. 
+#' @param genomicConstraint <GRanges>: GRanges object of constraint regions.
 #'  If NULL chromosomes in chromSizes are used as constraints (Default NULL)
-#' @param secondaryConst.var <character>: A string defining column name 
-#' containing compartment information in the metadata of anchor and 
+#' @param secondaryConst.var <character>: A string defining column name
+#' containing compartment information in the metadata of anchor and
 #' bait <GRanges> objects. (Default NULL)
-#' @param chromSizes <data.frame>: A data.frame containing chromosomes 
+#' @param chromSizes <data.frame>: A data.frame containing chromosomes
 #' names and lengths in base pairs.
-#' @param n_background <integer> : Number of background couples to keep. 
+#' @param n_background <integer> : Number of background couples to keep.
 #' (Default NULL)
-#' @param indexAnchor <GRanges>: A first indexed GRanges object used as 
+#' @param indexAnchor <GRanges>: A first indexed GRanges object used as
 #' pairs anchor (must be indexed using `IndexFeatures()`).
-#' @param indexBait <GRanges>: A second indexed GRanges object used as 
+#' @param indexBait <GRanges>: A second indexed GRanges object used as
 #' pairs bait (must be indexed using `IndexFeatures()`).
 #' @param cores <integer> : Number of cores used. (Default 1)
 #' @param areaFun <character or function>: A character or function that allows
-#'  to extract an area from each matrix that composes the matrices list 
+#'  to extract an area from each matrix that composes the matrices list
 #' (Default "center"). Look at [GetQuantif] for more info.
-#' @param operationFun <character or function>: A character or function 
-#' specifying the operation applied to the selected area (Default "mean"). 
+#' @param operationFun <character or function>: A character or function
+#' specifying the operation applied to the selected area (Default "mean").
 #' Look at [GetQuantif] for more info.
 #' @param bg_type <character>: Type of background couples to generate.
-#'\itemize{
+#' Possible choices: "random_anchors", "inter_TAD","inter_compartment",
+#' NULL (Defaults to "random_anchors"). More information in details...
+#' @param verbose <logical> details on progress? (Default: FALSE)
+#' @param z_score_per_dist_group <logical> Should z-scores and p.values
+#' for target couples be calculated in separate groups 
+#' based on distances? Look at details for more. (Default: TRUE)
+#' @param ... arguments to pass to [PrepareMtxList], inorder to treat
+#' background matrices.
+#' @return returns a <list> object with the z.test output for each
+#' target couple, values for the target couples and values for the
+#' background couples.
+#' 
+#' @export
+#' @importFrom GenomeInfoDb seqlevels
+#' @importFrom IRanges subsetByOverlaps
+#' @importFrom stats p.adjust sd pnorm
+#' @importFrom rlang .data
+#' @importFrom dplyr group_by mutate summarise left_join
+#' @details
+#' Types of background couples possible:
+#' \itemize{
 #' \item "random_anchors": picks random bins as anchors and forms couples 
 #' with bait bins.
 #' If genomicConstraint is supplied, only intra-TAD random-bait couples are
@@ -45,18 +65,17 @@
 #' different compartments.
 #' \item "NULL": If `NULL`, `random_anchors` are set by default.
 #' }
-#' @param verbose <logical> details on progress? (Default: FALSE)
-#' @param ... arguments to pass to [PrepareMtxList], inorder to treat 
-#' background matrices.
-#' @return returns a <list> object with the z.test output for each 
-#' target couple, values for the target couples and values for the 
-#' background couples.
-#' 
-#' @export
-#' @importFrom GenomeInfoDb seqlevels
-#' @importFrom IRanges subsetByOverlaps
-#' @importFrom stats p.adjust sd pnorm
-#' @importFrom rlang .data
+#'
+#' On the matter of `z_score_per_dist_group`:
+#' This option was included because we noticed that o/e values tend to be
+#' skewed towards very long distance interactions. As a result, long
+#' distance background couples tend to influence to mean and sd, resulting
+#' in only long distance target couples being significant.
+#' This option would allow the user to perform z-score computation on
+#' couples grouped into chunks of couples with comparative distances
+#' choosen as such: `c(0,50000 * 2^seq(1,15))`.
+#' This option is recommended when the target couples have a wide range
+#' of distance values.
 #' @examples
 #' h5_path <- system.file("extdata",
 #'     "Control_HIC_10k_2L.h5",
@@ -115,6 +134,7 @@ compare_to_background <- function(hicList = NULL,
                     bg_type = NULL,
                     cores = 1,
                     verbose = FALSE,
+                    z_score_per_dist_group = TRUE,
                     ...){
 
     targetCouples <- attributes(matrices)$interactions
@@ -333,7 +353,6 @@ compare_to_background <- function(hicList = NULL,
         operationFun = operationFun
     ) |> as.numeric() |> 
         `names<-`(attr(bg_counts,"names"))
-    # sigma.bg <- stats::sd(bg_quantifs)
 
     target_quantifs <- GetQuantif(
         matrices = matrices,
@@ -341,20 +360,41 @@ compare_to_background <- function(hicList = NULL,
         operationFun = operationFun
     ) |> as.numeric() |> 
         `names<-`(attr(matrices,"names"))
-    z_scores <- (target_quantifs - mean(bg_quantifs))- stats::sd(bg_quantifs)
-    pval_vector <- stats::pnorm(q = z_scores, lower.tail = FALSE)
-    z_output <- data.frame(names = names(target_quantifs),
-                    z.score = z_scores,
-                    p.value = pval_vector)
-    # z.output <- lapply(target_quantifs, 
-    #         function(i){broom::tidy(BSDA::z.test(x=bg_quantifs,
-    #                                         sigma.x = sigma.bg,
-    #                                         mu=i,
-    #                                         alternative = "less"))}
-    #         ) |> dplyr::bind_rows() |> 
-    #     dplyr::mutate(names = names(target_quantifs)) |>
-        # dplyr::mutate(adj.p = stats::p.adjust(p = .data$p.value,
-        #     method="bonferroni",n=length(target_quantifs)))
+    if(z_score_per_dist_group){
+        summary_bg_pergrp <- data.frame(
+        name = attributes(bg_counts)$names,
+        distance = attributes(bg_counts)$interactions$distance,
+        counts = bg_quantifs) |>
+        # chop background couples into separate groups with distance
+        dplyr::mutate(group=cut(.data$distance,
+                                breaks = c(0,50000 * 2^seq(1,15)),
+                                right = TRUE)) |>
+        dplyr::group_by(.data$group)|>
+        dplyr::summarise("avg" = mean(.data$counts),
+                        "sd" = stats::sd(.data$counts))
+
+        target_df <- data.frame(
+            name = attributes(matrices)$names,
+            distance = attributes(matrices)$interactions$distance,
+            counts = target_quantifs) |>
+            # chop target couples into separate groups with distance
+            dplyr::mutate(group = cut(.data$distance,
+                                        breaks = c(0,50000 * 2^seq(1,15)),
+                                        right = TRUE))
+        z_output <- dplyr::left_join(x = target_df, y = summary_bg_pergrp,
+                        by = 'group') |>
+                    dplyr::mutate(
+                        z.score = ((.data$counts-.data$avg)/.data$sd)) |>
+                    dplyr::mutate(
+                        p.value=pnorm(q = .data$z.score, lower.tail = FALSE))
+    } else {
+        z_scores <- (target_quantifs-mean(bg_quantifs))-stats::sd(bg_quantifs)
+        pval_vector <- stats::pnorm(q = z_scores, lower.tail = FALSE)
+        z_output <- data.frame(names = names(target_quantifs),
+                            z.score = z_scores,
+                            p.value = pval_vector)
+    }
+
     return(list(z.test = z_output, target_quantifs = target_quantifs,
         bg_quantifs = bg_quantifs))
 }